Move some util classes to iaf-commons module.
