+++ iaf-commons/src/main/java/nl/nn/adapterframework/functional/ThrowingRunnable.java
