+++ iaf-commons/src/main/java/nl/nn/adapterframework/util/StreamUtil.java
