+++ iaf-commons/src/main/java/nl/nn/adapterframework/util/StringDataSource.java
