+++ iaf-commons/src/main/java/nl/nn/adapterframework/util/StringUtil.java
