+++ iaf-commons/src/test/java/nl/nn/adapterframework/util/UUIDUtilTest.java
@@ -1,9 +1,10 @@
 package nl.nn.adapterframework.util;
 
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNotEquals;
 
-import static org.junit.jupiter.api.Assertions.*;
+import org.junit.jupiter.api.Test;
 
 class UUIDUtilTest {
