+++ iaf-commons/src/test/resources/StreamUtil/inISO8859-1.bin
