+++ iaf-commons/src/test/resources/StreamUtil/inUTF16BEwithBOM.bin
