+++ iaf-commons/src/test/resources/StreamUtil/inUTF16LEwithBOM.bin
