+++ iaf-commons/src/test/resources/StreamUtil/inUTF8noBOM.bin
