+++ iaf-commons/src/test/resources/StreamUtil/inUTF8withBOM.bin
