+++ iaf-commons/src/test/resources/StringResolver.properties
@@ -1,0 +1,14 @@
+key1=value1
+key2=${key1}
+key3=value2.${key1}
+key4=${key1}.${key3}
+key5=${key5}
+key1_value1=value101
+testMultiResolve=one,two,three_${key1}${key2},my_${key3},StageSpecifics_${key4}
+cyclic=prefix ${cyclic} suffix
+log.dir=c:/temp
+testdata=${log.dir}/testdata
+mfh=${testdata}/mfh
+dir=${path:-d:/temporary}/dir
+dir2=${log.dir:-c:/temporary}/dir
+allowedAliases=alias1,alias2
